\name{BEEM_EXPRS}
\alias{BEEM_EXPRS}
\docType{data}
\title{
A Log2 Gene Expression matrix
}
\description{
A Log2 Gene Expression matrix where rownames are genes and colnames are samples
}
\usage{data(BEEM_EXPRS)}
\format{
5,000 observations across 200 variables. All data is log2 transformed RSEM values. 
}
\keyword{datasets}
