\name{BEEM_MUT}
\alias{BEEM_MUT}
\docType{data}
\title{
A matrix containing discreet mutation calls
}
\description{
A matrix containing discreet mutation calls of either 'WT' or 'MUT' where rownames are genes and column names are samples
}
\usage{data(BEEM_MUT)}
\format{
5000 observations across 229 variables. 
}
\keyword{datasets}
