\name{BEEP_input_data}
\alias{BEEP_input_data}
\docType{data}
\title{
Gene by sample input data matrix.
}
\description{
Gene by sample input data matrix where rownames are genes and colnames are samples.  Scaled log2 gene expression data. 
}
\usage{data(BEEP_input_data)}
\format{
1,000 observations across 257 variables.
}
\keyword{datasets}
