\name{BEEP_sens_data}
\alias{BEEP_sens_data}
\docType{data}
\title{
Matrix containing sample names and discreet pharmacology calls.
}
\description{
A matrix containing 2 columns.  Column 1 contains sample names and must at least partially match column names of associated gene expression matrix.  Column 2 contains discreet pharmacology calls of which a subset should be labelled 'Resistant'.
}
\usage{data(BEEP_sens_data)}
\format{
257 observations across 2 variables.
}
\keyword{datasets}
