\name{BIGEE_example_input}
\alias{BIGEE_example_input}
\docType{data}
\title{
A data matrix containing log2 gene expression data
}
\description{
A data matrix containing log2 gene expression data where genes are rownames and samples are column names
}
\usage{data(BIGEE_example_input)}
\format{
1000 variables across 200 samples.  
}
\keyword{datasets}
