\name{BiSEp-package}
\alias{BiSEp-package}
\alias{BiSEp}
\docType{package}
\title{
BiSEp: Bimodality in gene expression to dissect tumours and reveal synthetic lethal drug targets and biomarkers
}
\description{
A set of tools that enable the user to accurately identify bimodality and non-normality in gene expression data and stratify samples as high or low expression for bimodal genes.  Enables identification of candidate synthetic lethal gene pairs. Enables the user to assess and visualise functional redundancy between candidate synthetic lethal gene pairs.
}
\details{
\tabular{ll}{
Package: \tab BiSEp\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2014-10-21\cr
License: \tab GPL-2\cr
}
This package has a mixture of CRAN and bioconductor packages listed as dependancies.  Please ensure that you have Bioconductor installed.
}
\author{
Author: Mark Wappett

Maintainer: Mark Wappett <mark.wappett@astrazeneca.com>

}

\keyword{ package }
