\name{BISEP_dat}
\alias{BISEP_dat}
\docType{data}
\title{
A list object containing 3 data frames.
}
\description{
Matrix 1 contains the output of the BISEP algorithm - including the midpoint of the bimodal distribution and the associated p value. Matrix 2 contains the output from the BI algorithm - including the delta, pi and BI values.  Matrix 3 contains the input matrix.
}
\usage{data(BISEP_dat)}
\format{
13 observations across 100 variables. 
}
\keyword{datasets}
