\name{MUT_data}
\alias{MUT_data}
\docType{data}
\title{
A matrix containing discreet mutation calls
}
\description{
A matrix containing discreet mutation calls of either 'WT' or 'MUT' where rownames are genes and column names are samples
}
\usage{data(MUT_data)}
\format{
4 observations across 442 variables. 
}
\keyword{datasets}
